==> step7_modelling_hybrid.ipynb <==
import pandas as pd
import numpy as np

from tqdm import tqdm

# Setup

from google.colab import drive

drive.mount('/content/drive')

Set your root directory below. Make sure the `/recommendations` folder is generated by the previous steps and is situated in this directory.

# Adjust your root directory
root = '/content/drive/MyDrive/KuaiRec/'

# Import the models' predictions

# root = '../'

test_pred_ncf = pd.read_csv(root + 'recommendations/final_w_clustering_batch_size512_num_epochs20_lr0_001_embedding.csv')
test_pred_caption = pd.read_csv(root + 'recommendations/recommendations_caption_test_full.csv')

# Combine the two predictions
The weights are gotten from the results of a regression analysis. It can be found on `step_6_regression.ipynb`

weight_ncf = 0.5552
weight_caption = 0.1421

caption_ratio_dict = test_pred_caption.set_index(['user_id', 'video_id'])['watch_ratio'].to_dict()

hybrid_pred = test_pred_ncf.copy()

for i in tqdm(range(len(hybrid_pred))):
    user_id = hybrid_pred['user_id'][i]
    video_id = hybrid_pred['video_id'][i]
    ncf_ratio = hybrid_pred['watch_ratio'][i]
    caption_ratio = 0
    if (user_id, video_id) in caption_ratio_dict:
        caption_ratio = caption_ratio_dict[(user_id, video_id)]
    hybrid_pred.at[i, 'watch_ratio'] = ncf_ratio * weight_ncf + caption_ratio * weight_caption

hybrid_pred.to_csv(root + 'recommendations/recommendations_hybrid_test_full.csv', index=False)
