==> step8_evaluation.ipynb <==
# Evaluation

## Set up

from google.colab import drive

drive.mount('/content/drive')

Set your root directory below. Make sure the `/data`, `/data_exports` and `/recommendations` folders are uploaded and is situated in this directory.

# Adjust your root directory
root = '/content/drive/MyDrive/KuaiRec/'

## Load Data and Predictions

import pandas as pd

import eval_fns

In total, we have eight sets of predictions - 2 per model. One set is generated from only the training set and is to be evaluated against the validation set, while the other is generated after training on both the training and validation sets and to be evaluated on the final testing set. Hyperparameters for the models were chosen based on the validation set, and the final predictions are generated using these hyperparameters. The test set represents the final unseen data, and the performance of the models on this set is the most important.

There is also an additional set of predictions from the NCF model, trained on the train and validation data, without user segmentation. This will be evaluated on the test dataset and compared to the performance of the NCF model with user segmentation. With this analysis, we want to see how much does the user segmentation contribute to the NCF model overall performance.

root = './'

# Predictions to be tested on the validation set
prediction_scores_caption_on_val = pd.read_csv(root + 'recommendations/recommendations_caption_val_full.csv')
prediction_scores_ncf_on_val = pd.read_csv(root + 'recommendations/w_clustering_batch_size512_num_epochs20_lr0.001_embedding_dim64_dropout0.3_decay0.01.csv')
prediction_scores_hybrid_on_val = pd.read_csv(root + 'recommendations/recommendations_hybrid_val_full.csv')
prediction_scores_random_on_val = pd.read_csv(root + 'recommendations/recommendations_random_val_full.csv')

# Predictions to be tested on the test set
prediction_scores_caption_on_test = pd.read_csv(root + 'recommendations/recommendations_caption_test_full.csv')
prediction_scores_ncf_on_test = pd.read_csv(root + 'recommendations/final_w_clustering_batch_size512_num_epochs20_lr0.001_embedding_dim64_dropout0.3_decay0.01.csv')
prediction_scores_hybrid_on_test = pd.read_csv(root + 'recommendations/recommendations_hybrid_test_full.csv')
prediction_scores_random_on_test = pd.read_csv(root + 'recommendations/recommendations_random_test_full.csv')

prediction_scores_ncf_on_test_without_clustering = pd.read_csv(root + 'recommendations/final_wo_clustering_batch_size512_num_epochs20_lr0.001_embedding_dim64_dropout0.3_decay0.01.csv')

# Obtain the ground truth watch ratios from data
joined_train_data = pd.read_csv(root + 'data_exports/joined_train_data.csv')
joined_val_data = pd.read_csv(root + 'data_exports/joined_val_data.csv')
joined_test_data = pd.read_csv(root + 'data_exports/joined_test_data.csv')

joined_train_val_data = pd.concat([joined_train_data, joined_val_data])

# Load the video data in order to get the video categories
video_data = pd.read_csv(root + 'data/kuairec_caption_category_translated.csv', index_col=0)

## Data Preprocessing

# Rename
prediction_scores_caption_on_val = prediction_scores_caption_on_val.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_ncf_on_val = prediction_scores_ncf_on_val.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_hybrid_on_val = prediction_scores_hybrid_on_val.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_random_on_val = prediction_scores_random_on_val.rename(columns={'watch_ratio': 'predicted_watch_ratio'})

prediction_scores_caption_on_test = prediction_scores_caption_on_test.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_ncf_on_test = prediction_scores_ncf_on_test.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_hybrid_on_test = prediction_scores_hybrid_on_test.rename(columns={'watch_ratio': 'predicted_watch_ratio'})
prediction_scores_random_on_test = prediction_scores_random_on_test.rename(columns={'watch_ratio': 'predicted_watch_ratio'})

prediction_scores_ncf_on_test_without_clustering = prediction_scores_ncf_on_test_without_clustering.rename(columns={'watch_ratio': 'predicted_watch_ratio'})

# Sort
prediction_scores_caption_on_val = prediction_scores_caption_on_val.sort_values(by=['user_id', 'video_id'])
prediction_scores_ncf_on_val = prediction_scores_ncf_on_val.sort_values(by=['user_id', 'video_id'])
prediction_scores_hybrid_on_val = prediction_scores_hybrid_on_val.sort_values(by=['user_id', 'video_id'])
prediction_scores_random_on_val = prediction_scores_random_on_val.sort_values(by=['user_id', 'video_id'])

prediction_scores_caption_on_test = prediction_scores_caption_on_test.sort_values(by=['user_id', 'video_id'])
prediction_scores_ncf_on_test = prediction_scores_ncf_on_test.sort_values(by=['user_id', 'video_id'])
prediction_scores_hybrid_on_test = prediction_scores_hybrid_on_test.sort_values(by=['user_id', 'video_id'])
prediction_scores_random_on_test = prediction_scores_random_on_test.sort_values(by=['user_id', 'video_id'])

prediction_scores_ncf_on_test_without_clustering = prediction_scores_ncf_on_test_without_clustering.sort_values(by=['user_id', 'video_id'])

## Get user watch history

We want to be able to filter out videos that the user has already watched. This is so that we recommend new videos instead.

user_watch_history_from_train = eval_fns.get_user_watch_history(joined_train_data)
user_watch_history_from_train_val = eval_fns.get_user_watch_history(joined_train_val_data)

## Getting ground truth videos for each user

Next, we process the test set to obtain the ground truth watch ratios. The test set is filtered to only contain videos that are present in training and validation data, as well as those that the user has not watched before. Users and videos that are not in the training data are filtered out as well, as we cannot make recommendations for them. The remaining data is then sorted by user in ascending order and watch_ratio in descending order.

# Obtain users and videos in training data
users_in_train_data = set(joined_train_data['user_id'])
videos_in_train_data = set(joined_train_data['video_id'])

# Obtain users and videos in training and validation data
users_in_train_val_data = set(joined_train_val_data['user_id'])
videos_in_train_val_data = set(joined_train_val_data['video_id'])

# Get ground truth user-item combinations and their watch ratios
ground_truth_val = eval_fns.get_ground_truth(joined_val_data[['user_id', 'video_id', 'watch_ratio']], users_in_train_data, videos_in_train_data, user_watch_history_from_train)
ground_truth_test = eval_fns.get_ground_truth(joined_test_data[['user_id', 'video_id', 'watch_ratio']], users_in_train_val_data, videos_in_train_val_data, user_watch_history_from_train_val)

# Ground ground truth watch ratios for user 14
ground_truth_test[ground_truth_test['user_id'] == 14]

## Getting recommendations for each user

With the prediction scores generated from our models, we obtain the video recommendations for each user. This is done by first filtering for videos that the user has not watched before, then sorting the predicted watch ratio in descending order.

# Get recommendations on validation set
videos_in_val_data = set(joined_val_data['video_id'])

recommendations_caption_for_val = eval_fns.get_user_recommendations(prediction_scores_caption_on_val, videos_in_val_data, user_watch_history_from_train)
recommendations_ncf_for_val = eval_fns.get_user_recommendations(prediction_scores_ncf_on_val, videos_in_val_data, user_watch_history_from_train)
recommendations_hybrid_for_val = eval_fns.get_user_recommendations(prediction_scores_hybrid_on_val, videos_in_val_data, user_watch_history_from_train)
recommendations_random_for_val = eval_fns.get_user_recommendations(prediction_scores_random_on_val, videos_in_val_data, user_watch_history_from_train)

# Get recommendations on test set
videos_in_test_data = set(joined_test_data['video_id'])

recommendations_caption_for_test = eval_fns.get_user_recommendations(prediction_scores_caption_on_test, videos_in_test_data, user_watch_history_from_train_val)
recommendations_ncf_for_test = eval_fns.get_user_recommendations(prediction_scores_ncf_on_test, videos_in_test_data, user_watch_history_from_train_val)
recommendations_hybrid_for_test = eval_fns.get_user_recommendations(prediction_scores_hybrid_on_test, videos_in_test_data, user_watch_history_from_train_val)
recommendations_random_for_test = eval_fns.get_user_recommendations(prediction_scores_random_on_test, videos_in_test_data, user_watch_history_from_train_val)

recommendations_ncf_for_test_without_clustering = eval_fns.get_user_recommendations(prediction_scores_ncf_on_test_without_clustering, videos_in_test_data, user_watch_history_from_train_val)

# Recommendations from NCF model for user 14 to be evaluated on the test set
recommendations_ncf_for_test[recommendations_ncf_for_test['user_id'] == 14]

## Evaluation Metrics

We have chosen various evaluation metrics to provide a comprehensive evaluation of our models' performance. They can be grouped into 3 broad categories - Engagement, Relevance and Diversity.

### Engagement
1. **Average Watch Ratio @ k**: Measures the average proportion of content that users watch across all recommended videos.

### Relevance
1. **Precision@k**: Proportion of recommended videos in top K that are relevant.

2. **Recall@k**: Proportion of all relevant videos that appear in top K recommendations.

3. **F1-Score@k**: The harmonic mean of precision and recall at K, balancing the trade-off between recommending relevant videos (precision) and capturing all relevant videos (recall). 

As these metrics require a binary label, we establish a threshold for predicted_watch_ratio of 0.7, where if a video has `predicted_watch_ratio` >= 0.7: relevant, `predicted_watch_ratio` < 0.7: irrelevant.

### Diversity
1. **Category-Aware NDCG @ k**: Measures how well the recommended videos' category distribution matches the user's true category preference ranking.

2. **Distinct Categories @ k**: Number of distinct categories that appear in the top K recommendations.


We have chosen k to be 100.

k = 100
threshold = 0.5

# For Validation
reco_grp_caption_for_val = recommendations_caption_for_val.groupby('user_id')
reco_grp_ncf_for_val = recommendations_ncf_for_val.groupby('user_id')
reco_grp_hybrid_for_val = recommendations_hybrid_for_val.groupby('user_id')
reco_grp_random_for_val = recommendations_random_for_val.groupby('user_id')

ground_truth_grp_for_val = ground_truth_val.groupby('user_id')

# For Test
reco_grp_caption_for_test = recommendations_caption_for_test.groupby('user_id')
reco_grp_ncf_for_test = recommendations_ncf_for_test.groupby('user_id')
reco_grp_hybrid_for_test = recommendations_hybrid_for_test.groupby('user_id')
reco_grp_random_for_test = recommendations_random_for_test.groupby('user_id')

reco_grp_ncf_for_test_without_clustering = recommendations_ncf_for_test_without_clustering.groupby('user_id')

ground_truth_grp_for_test = ground_truth_test.groupby('user_id')

## Performance Across Models on the Validation Set
### Results

metrics_df_caption_for_val = eval_fns.get_all_metrics(k, ground_truth_grp_for_val, reco_grp_caption_for_val, video_data, threshold, by_cluster = False)
metrics_df_ncf_for_val = eval_fns.get_all_metrics(k, ground_truth_grp_for_val, reco_grp_ncf_for_val, video_data, threshold, by_cluster = False)
metrics_df_hybrid_for_val = eval_fns.get_all_metrics(k, ground_truth_grp_for_val, reco_grp_hybrid_for_val, video_data, threshold, by_cluster = False)
metrics_df_random_for_val = eval_fns.get_all_metrics(k, ground_truth_grp_for_val, reco_grp_random_for_val, video_data, threshold, by_cluster = False)

# Concatenate the metrics dataframes
metrics_for_val_combined = pd.concat([metrics_df_ncf_for_val, metrics_df_caption_for_val, metrics_df_hybrid_for_val, metrics_df_random_for_val], axis=0)

# Add model names
metrics_for_val_combined.index = ['Neural Collaborative Filtering with Time Decay', 'Caption-based Video Filtering with Time Decay', 'Hybrid', 'Random']

metrics_for_val_combined.drop(columns=['cluster'], inplace=True)

metrics_for_val_combined

## Performance Across Models on the Testing Set
### Results

metrics_df_caption_for_test = eval_fns.get_all_metrics(k, ground_truth_grp_for_test, reco_grp_caption_for_test, video_data, threshold, by_cluster = False)
metrics_df_ncf_for_test = eval_fns.get_all_metrics(k, ground_truth_grp_for_test, reco_grp_ncf_for_test, video_data, threshold, by_cluster = False)
metrics_df_hybrid_for_test = eval_fns.get_all_metrics(k, ground_truth_grp_for_test, reco_grp_hybrid_for_test, video_data, threshold, by_cluster = False)
metrics_df_random_for_test = eval_fns.get_all_metrics(k, ground_truth_grp_for_test, reco_grp_random_for_test, video_data, threshold, by_cluster = False)

metrics_df_ncf_for_test_without_clustering = eval_fns.get_all_metrics(k, ground_truth_grp_for_test, reco_grp_ncf_for_test_without_clustering, video_data, threshold, by_cluster = False)

# Concatenate the metrics dataframes
metrics_for_test_combined = pd.concat([metrics_df_ncf_for_test, metrics_df_caption_for_test, metrics_df_hybrid_for_test, metrics_df_random_for_test], axis=0)

# Add model names
metrics_for_test_combined.index = ['Neural Collaborative Filtering with Time Decay', 'Caption-based Video Filtering with Time Decay', 'Hybrid', 'Random']

metrics_for_test_combined.drop(columns=['cluster'], inplace=True)

metrics_for_test_combined

### Insights
1. Engagement

    As we can see, the Average Watch Ratio @ k is higher for all of our models compared to the random baseline, indicating that the users are watching more of the recommended videos. This is a good sign as it shows that our models are able to recommend videos that users are more likely to watch and enjoy. The NCF model performs much better than the Caption-based model, which is expected as the NCF model is a much more complex model that is able to capture more complex patterns in the data.

2. Relevance

    The Precision@k, Recall@k and F1-Score@k are all higher for all of our models compared to the random baseline. This indicates that our models are able to recommend more relevant videos to the users. The higher F1-score@k also indicates that our models are able to balance the trade-off between recommending relevant videos and capturing all relevant videos. Similar to the engagement metrics, the NCF model performs much better than the Caption-based model.

3. Diversity

    In terms of Distinct Categories @ k, we can see that all our models recommend fewer number of distinct categories compared to the random baseline. Caption-based model recommends the least number of distinct categories, which is expected as the model's recommendations were partially based on the embeddings of the category of the video. In general, this suggests that our models are more targeted than random recommendations, but might not be as diverse in terms of the categories of the recommended videos. However, as we can see from the Category-Aware NDCG @ 100 metric, the NCF model has higher scores compared to our baseline, which means that even though the number of distinct categories recommended is lower, the categories of the recommended videos are more aligned with the user's true category preference ranking.

### NCF Model Performance

Previously, we segmented users into four distinct clusters based on their behavioral patterns, in hopes to capture subtle patterns unique to each group and improve model performance. Hence, we have trained the NCF model on both segmented and non-segmented combined train and validation data. 

Let us see if performance is indeed better with customer segmentation. 

#### Performance With User Segmentation

metrics_ncf_per_cluster = eval_fns.get_all_metrics(k, ground_truth_test, recommendations_ncf_for_test, video_data, threshold, by_cluster=True)

metrics_ncf_per_cluster

#### Performance Without Segmentation

metrics_ncf_per_without_clustering = eval_fns.get_all_metrics(k, ground_truth_test, recommendations_ncf_for_test_without_clustering, video_data, threshold, by_cluster=False)

metrics_ncf_per_without_clustering

We can see that the recommendations made with user segmentation outperforms recommendations made without user segmentation on all metrics. Hence, user segmentation is indeed able to provide more tailored recommendations to users.
